Overall WebApp files uploaded
This webapp was built as a college project, with that said I didn't even had the opportunity to create a repository for it sooner. This is a commit of the full app files that make it run and work in every mahcine, including SQL file to replicate the DB, and backend files for the NodeJS API that is consumed by Angular app.

This was supposed to be just a prototype, so the UI/UX isn't the prettiest but the funcitonality of all the requirements asked are there. After the semester is finished I will get back to it to fully finish it, and by that I mean finishing the JWT interceptor which I didn't have time to implement, auth guards, better UI and overall functionallity. 

I will be commiting the updates to this repo.